+++ fixthis-gradle-plugin/src/main/kotlin/io/beyondwin/fixthis/gradle/FixThisExtension.kt
@@ -11,7 +11,7 @@ abstract class FixThisExtension @Inject constructor(
         objects.property(Boolean::class.java).convention(true)
 
     val runtimeVersion: Property<String> =
-        objects.property(String::class.java).convention("0.1.0")
+        objects.property(String::class.java).convention(FixThisPluginVersion.defaultRuntimeVersion())
 
     val addDebugRuntime: Property<Boolean> =
         objects.property(Boolean::class.java).convention(true)